\name{Sample250}
\alias{Sample250}
\title{Sample250}
\description{
  Sample250.txt includes the first 1000 lines in SupplementaryTable2.txt which is a 
  supplementary file of RNAseq for Marioni,J.C. et al. (2008) (\url{http://genome.cshlp.org/content/18/9/1509/suppl/DC1}). 
}
\references{
  Marioni,J.C. et al. (2008) RNA-seq: an assessment of technical reproducibility and comparison with
  gene expression arrays. \emph{Genome Res.}, \bold{18}, 1509-1517.
}
\seealso{
 \code{\link{PermuteDEG}},
}
\keyword{datasets}
